+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/ui/search/RepositorySearchFragment.kt
@@ -13,7 +13,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import jp.co.yumemi.android.codecheck.data.model.RepositoryItem
-import jp.co.yumemi.android.codecheck.ui.theme.MainTheme
+import jp.co.yumemi.android.codecheck.designsystem.MainTheme
 
 @AndroidEntryPoint
 class RepositorySearchFragment : Fragment() {
