+++ core/designsystem/src/main/java/jp/co/yumemi/android/codecheck/designsystem/Color.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codecheck.ui.theme
+package jp.co.yumemi.android.codecheck.designsystem
 
 import androidx.compose.ui.graphics.Color
